requirements.txt
amqp==5.2.0
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
billiard==4.2.0
blinker==1.7.0
boto3==1.34.91
botocore==1.34.91
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
filelock==3.13.4
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parsel==1.9.1
prompt-toolkit==3.0.43
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
queuelib==1.6.2
redis==5.0.4
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.1
scraperapi-sdk==0.2.2
Scrapy==2.11.1
scrapy-splash==0.9.0
sentry-sdk==1.45.0
service-identity==24.1.0
six==1.16.0
SQLAlchemy==2.0.29
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.2
zope.interface==6.3
